🩹 Fixed a toString-dependent storage path, and retracted a false review claim
The follow-up review confirmed the load-bearing things held: spec.md and the
ticket files were untouched, so nothing was bent to close a finding; parallelism
is genuinely still parallel; and the never-under-reports property test plus the
test pinning open item 18.6 #4 are byte-identical.

It also found that uriPrefix had quietly come to depend on GroupId.toString().
"$bucket/snapshots/$group/..." produced the right string only because GroupId
overrides toString() = value, which no test pins - so deleting that override
would have silently rewritten every object key in the bucket while the suite
stayed green. Now group.value, like the eight binds beside it.

The more important correction is documentary. The previous entry claimed each
consolidated fake was a strict superset. That is false for the object store in
the direction that matters: ArchiverTest's fake overrode only put and sizeOf, so
anything else fell through to a client pointed at 127.0.0.1:1 and failed loudly.
That refusal was a capability - if Archiver ever started deleting objects, that
suite used to notice, and now it does not. A weakening of an earlier phase's
test, asserted away rather than seen.

Two questions are left open rather than self-approved a third time: whether the
fixture consolidation should have edited three earlier-phase suites at all, and
whether restoring the store's refusal is worth another edit to them. In one
commit the size-budget rule was escalated correctly while this one was decided
unilaterally; the same rule should not get whichever reading is convenient.

200 tests, 0 failures.

+++ snapshotcache/src/test/kotlin/infra/snapshotarchive/ArchiveMaintenanceTest.kt
@@ -1,12 +1,6 @@
 package infra.snapshotarchive
 
-import infra.snapshotcache.api.CopyOutResult
-import infra.snapshotcache.api.CopyOutSpec
-import infra.snapshotcache.api.GenerationInfo
 import infra.snapshotcache.api.GroupId
-import infra.snapshotcache.api.Snapshot
-import infra.snapshotcache.api.SnapshotCache
-import io.minio.MinioClient
 import org.assertj.core.api.Assertions.assertThat
 import org.jdbi.v3.core.Jdbi
 import org.junit.jupiter.api.AfterEach
@@ -19,23 +13,17 @@ import org.testcontainers.junit.jupiter.Testcontainers
 import org.testcontainers.oracle.OracleContainer
 import java.nio.file.Files
 import java.nio.file.Path
-import java.sql.Connection
 import java.sql.DriverManager
 import java.time.Clock
 import java.time.Duration
 import java.time.Instant
 import java.time.ZoneOffset
-import java.util.concurrent.ConcurrentHashMap
 import java.util.concurrent.CopyOnWriteArrayList
 import java.util.concurrent.CountDownLatch
 import java.util.concurrent.TimeUnit
 import java.util.concurrent.atomic.AtomicInteger
 import java.util.concurrent.atomic.AtomicLong
 import java.util.concurrent.atomic.AtomicReference
-import java.util.logging.Handler
-import java.util.logging.Level
-import java.util.logging.LogRecord
-import java.util.logging.SimpleFormatter
 
 /**
  * P13 acceptance: convergence and retention.
@@ -302,7 +290,7 @@ class ArchiveMaintenanceTest {
     ): ManifestEntry {
         val inventory = tables.map { ArchivedObject(it, "$it.parquet", BYTES, CHECKSUM, 10) }
         val entry = dao.insertPending(group, dataAsOf, Inventory.encode(inventory), generation = 1)
-        upload.forEach { store.seed(keyOf(entry, "$it.parquet"), BYTES) }
+        upload.forEach { store.seed(keyOf(entry, "$it.parquet"), BYTES.toInt()) }
         return entry
     }
